Add files via upload
This is the full code to run this program. The Actor class and SixDegrees class were my creation. The Graph class was edited for my purpose.
I am not the author of the  LinkedList, Queue, and Stack classes, as well as NodeType.h, actors.txt and expected_output.txt. All other files are written by me.